requirements.txt
sqlmodel==0.0.14
fastapi==0.111.0
asgiref==3.7.2
pydantic-settings==2.1.0
pydantic==2.5.2
fastapi-limiter==0.1.5
python-dotenv==1.0.1
loguru==0.7.2
starlette==0.37.2
uvicorn==0.24.0.post1
redis==4.6.0
asyncmy==0.2.9
pyjwt==2.8.0
pandas~=2.2.2
sqlalchemy~=2.0.23
motor~=3.5.1
funboost~=46.2
pytz~=2020.5
fire~=0.6.0
